+++ packages/create-rainbowkit/generated-test-app/CHANGELOG.md
@@ -1,5 +1,15 @@
 # generated-test-app
 
+## 0.1.60
+
+### Patch Changes
+
+- Updated dependencies [08e3f4c]
+- Updated dependencies [cb3614e]
+- Updated dependencies [53d96bc]
+- Updated dependencies [bfab830]
+  - @rainbow-me/rainbowkit@1.0.5
+
 ## 0.1.59
 
 ### Patch Changes